api/pyproject.toml
[tool.poetry]
name = "slow-api"
version = "v0.1.34"
description = "banger."
authors = ["derpinskV2 <kderpinski@proton.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
argon2-cffi = "^23.1.0"
celery = "^5.4.0"
channels = "^4.2.0"
channels-redis = "^4.2.1"
django = "^5.1.4"
django-celery-results = "^2.5.1"
django-cors-headers = "^4.6.0"
django-environ = "^0.11.2"
django-ninja = "^1.3.0"
django-ninja-extra = "^0.22.0"
django-ninja-jwt = {extras = ["crypto"], version = "^5.3.5"}
django-redis = "^5.4.0"
orjson = "^3.10.13"
pandas = "^2.2.3"
pre-commit = "^4.0.1"
psycopg = {extras = ["binary"], version = "^3.2.3"}
pydantic = {extras = ["email"], version = "^2.10.4"}
python-slugify = "^8.0.4"
qrcode = {extras = ["pil"], version = "^8.0"}
rich = "^13.9.4"
setuptools = "^75.7.0"
toml = "^0.10.2"
urllib3 = "^2.3.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
wheel = "^0.45.1"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^24.10.0"}
django-extensions = "^3.2.3"
faker = "^33.3.0"
model-bakery = "^1.20.1"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
